adding name atribute to process class

+++ fcfs.cpp
@@ -8,6 +8,6 @@ void FCFS::setParameters(QVector<Process> process){
 
 }
 
-void FCFS::getResults(QString *ToDrawSignal, QString *ToDrawProcess, int *averageWaitingTime){
+void FCFS::getResults(QString *ToDrawSignal, QString *ToDrawProcess, float *averageWaitingTime){
 
 }
