+++ priority.h
@@ -9,7 +9,7 @@ class Priority
 public:
     Priority();
     void setParameters(QVector<Process> process,bool preemptiveOrNot,int QuantumTime);
-    void getResults(QString *ToDrawSignal, QString *ToDrawProcess, int* averageWaitingTime);
+    void getResults(QString *ToDrawSignal, QString *ToDrawProcess, float* averageWaitingTime);
 };
 
 #endif // PRIORITY_H
