+++ sjf.cpp
@@ -9,6 +9,6 @@ void SJF::setParameters(QVector<Process> process, bool preemptiveOrNot){
 
 }
 
-void SJF::getResults(QString *ToDrawSignal, QString *ToDrawProcess, int *averageWaitingTime){
+void SJF::getResults(QString *ToDrawSignal, QString *ToDrawProcess, float *averageWaitingTime){
 
 }
