+++ sjf.h
@@ -9,7 +9,7 @@ class SJF
 public:
     SJF();
     void setParameters(QVector<Process> process,bool preemptiveOrNot);
-    void getResults(QString *ToDrawSignal, QString *ToDrawProcess, int* averageWaitingTime);
+    void getResults(QString *ToDrawSignal, QString *ToDrawProcess, float* averageWaitingTime);
 };
 
 #endif // SJF_H
